requirements.txt
appium-python-client==3.1.0
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-pytest==2.13.2
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
